+++ packages/amplify-ui-components/src/components/amplify-sign-out/amplify-sign-out.tsx
@@ -14,7 +14,7 @@ export class AmplifySignOut {
   /** Auth state change handler for this component */
   @Prop() handleAuthStateChange: AuthStateHandler = dispatchAuthStateChangeEvent;
   /** Text inside of the Sign Out button */
-  @Prop() buttonText: string = I18n.get(Translations.SIGN_OUT);
+  @Prop() buttonText: string = Translations.SIGN_OUT;
 
   private async signOut(event) {
     if (event) event.preventDefault();
@@ -36,7 +36,7 @@ export class AmplifySignOut {
   render() {
     return (
       <amplify-button slot="sign-out" onClick={event => this.signOut(event)} data-test="sign-out-button">
-        {this.buttonText}
+        {I18n.get(this.buttonText)}
       </amplify-button>
     );
   }